+++ src/editor/actions/sagas/fetch-cell-saga.js
@@ -1,9 +1,7 @@
 import { all, call, put } from "redux-saga/effects";
 
-import generateRandomId from "../../../shared/utils/generate-random-id";
 import parseFetchCell from "../fetch-cell-parser";
 import { addToConsoleHistory } from "../console-message-actions";
-import { loadFile } from "../file-request-actions";
 import { evaluateLanguagePlugin } from "./language-plugin-saga";
 
 import { triggerEvalFrameTask } from "./eval-frame-sender";
@@ -14,6 +12,7 @@ import {
   successMessage,
   syntaxErrorToString
 } from "../../../shared/utils/fetch-tools";
+import { loadFileFromServer } from "../../../shared/utils/file-operations";
 
 function fetchProgressInitialStrings(fetchInfo) {
   let text;
@@ -29,22 +28,9 @@ function fetchProgressInitialStrings(fetchInfo) {
 }
 
 function* handleValidFetch(fetchInfo, historyId, lineIndex) {
-  const extractFileNameFromLocalFilePath = filepath =>
-    filepath
-      .split("files/")
-      .slice(1)
-      .join("");
-
   const { filePath, fetchType, isRelPath } = fetchInfo.parsed;
-  const fileFetcher = isRelPath
-    ? filepath => {
-        loadFile(
-          extractFileNameFromLocalFilePath(filepath),
-          generateRandomId(),
-          fetchType
-        );
-      }
-    : genericFetch;
+  const fileFetcher = isRelPath ? loadFileFromServer : genericFetch;
+  console.log("FETCH INFO ------------------", filePath, fetchInfo);
 
   let fetchedFile;
   try {
@@ -56,8 +42,15 @@ function* handleValidFetch(fetchInfo, historyId, lineIndex) {
       lineIndex,
       lineContent: errorMessage(fetchInfo, err).text
     });
-    return errorMessage(fetchInfo, err);
-    // throw new Error(`File fetch failed; throwing to halt eval queue`);
+    yield put({
+      type: "UPDATE_VALUE_IN_HISTORY",
+      historyItem: {
+        historyId,
+        level: "ERROR" // hasError ? "ERROR" : undefined
+      }
+    });
+    throw new Error(`failed to fetch file; halting eval queue`);
+    // return errorMessage(fetchInfo, err);
   }
 
   if (["text", "json", "blob", "arrayBuffer"].includes(fetchType)) {
@@ -84,18 +77,8 @@ function* handleValidFetch(fetchInfo, historyId, lineIndex) {
     lineIndex,
     lineContent: successMessage(fetchInfo).text
   });
-  // yield put(
-  //   addToConsoleHistory({
-  //     historyType: "FETCH_CELL_INFO",
-  //     content: progressStrings.map(t => t.text).join("")
-  //   })
-  // );
-  // console.log(
-  //   "successMessage(fetchInfo) ----------- ",
-  //   successMessage(fetchInfo),
-  //   fetchInfo
-  // );
-  return successMessage(fetchInfo);
+
+  // return successMessage(fetchInfo);
 }
 
 export function* evaluateFetch(fetchText) {
@@ -106,44 +89,40 @@ export function* evaluateFetch(fetchText) {
       addToConsoleHistory({
         historyType: "FETCH_CELL_INFO",
         content: syntaxErrors.map(fetchProgressInitialStrings).map(t => t.text),
-        // .join("\n")
         level: "ERROR"
       })
     );
 
-    throw new Error(
-      `Fetch cell has syntax errors; throwing to halt eval queue`
-    );
+    throw new Error(`Syntax errors in fetch chunk; halting eval queue`);
   }
 
   const progressStrings = fetches.map(fetchProgressInitialStrings);
   const { historyId } = yield put(
     addToConsoleHistory({
       historyType: "FETCH_CELL_INFO",
-      content: progressStrings.map(t => t.text) // .join("\n")
+      content: progressStrings.map(t => t.text)
     })
   );
 
-  const outcomes = yield all(
+  // const outcomes =
+  yield all(
     fetches.map((fetchSpec, i) =>
       call(handleValidFetch, fetchSpec, historyId, i)
     )
   );
 
-  const errors = outcomes.filter(f => f.text.startsWith("ERROR"));
-  const hasError = errors.length > 0;
+  // const errors = outcomes.filter(f => f.text.startsWith("ERROR"));
+  // const hasError = errors.length > 0;
 
-  yield put({
-    type: "UPDATE_VALUE_IN_HISTORY",
-    historyItem: {
-      historyId,
-      // content: outcomes.map(t => t.text),
-      level: hasError ? "ERROR" : undefined
-    }
-  });
+  // yield put({
+  //   type: "UPDATE_VALUE_IN_HISTORY",
+  //   historyItem: {
+  //     historyId,
+  //     level: hasError ? "ERROR" : undefined
+  //   }
+  // });
 
-  if (hasError) {
-    throw new Error(`Fetch outcome error; throwing to halt eval queue`);
-  }
-  console.log("FETCH completed ok >>>>>>>>>>>>>>>>>>");
+  // if (hasError) {
+  //   throw new Error(`Fetch outcome error; halting eval queue`);
+  // }
 }