+++ src/eval-frame/components/panes/history-item.jsx
@@ -70,7 +70,6 @@ export class HistoryItemUnconnected extends React.Component {
 }
 
 export function mapStateToProps(state, ownProps) {
-  // return state.history.filter(h => h.historyId === ownProps.historyId)[0];
   const historyItem = state.history.filter(
     h => h.historyId === ownProps.historyId
   )[0];